chore: automatically generate data types from schemas
