+++ package.json
@@ -2,7 +2,8 @@
   "private": true,
   "scripts": {
     "dev": "next dev",
-    "build": "next build && next export",
+    "build": "npm run compile:types && next build && next export",
+    "compile:types": "ts-node --skipProject scripts/compile-types.ts",
     "start": "next start"
   },
   "dependencies": {
@@ -12,6 +13,7 @@
     "@types/react-transition-group": "^4.4.5",
     "autoprefixer": "^10.4.13",
     "eslint-config-next": "^13.1.6",
+    "json-schema-to-typescript": "^11.0.3",
     "next": "^13.1.6",
     "postcss": "^8.4.21",
     "prettier": "^2.8.3",
@@ -22,6 +24,7 @@
     "react-pro-sidebar": "^1.0.0",
     "react-transition-group": "^4.4.5",
     "tailwindcss": "^3.2.6",
+    "ts-node": "^10.9.1",
     "typescript": "^4.9.5",
     "yaml": "^2.2.1"
   }
